+++ src/sqlite/teams/permission.rs
@@ -51,14 +51,14 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn get_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
     ) -> Result<PermissionSet<R, A>, AuthError> {
         let row: Option<PermissionRecord> = sqlx::query_as(
             "SELECT permissions FROM team_member_permissions WHERE team_id = ? AND user_id = ?",
         )
-        .bind(team_id)
-        .bind(user_id)
+        .bind(team_id as i64)
+        .bind(user_id as i64)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -78,8 +78,8 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self, permissions), err))]
     async fn set_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         permissions: &PermissionSet<R, A>,
     ) -> Result<(), AuthError> {
         let json_str = permissions.to_json();
@@ -93,8 +93,8 @@ where
             DO UPDATE SET permissions = ?, updated_at = ?
             ",
         )
-        .bind(team_id)
-        .bind(user_id)
+        .bind(team_id as i64)
+        .bind(user_id as i64)
         .bind(&json_str)
         .bind(&json_str)
         .bind(now)
@@ -114,8 +114,8 @@ where
     )]
     async fn grant_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: R,
         action: A,
     ) -> Result<(), AuthError> {
@@ -130,8 +130,8 @@ where
     )]
     async fn revoke_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: &R,
         action: &A,
     ) -> Result<(), AuthError> {
@@ -146,8 +146,8 @@ where
     )]
     async fn has_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: &R,
         action: &A,
     ) -> Result<bool, AuthError> {
